The main change is maintainace of tensorboard as a cli tool, the two minor ones: 1. logger dir has been removed, no it's recommended to use ... 2>&1 | tee log.txt, 2. desirable structure of storage has been changed: cv_fall_2022/lesson_6_nn_intro/experiments -> experiments/cv_fall_2022/lesson_6_nn_intro

+++ cli/train/metric_factory/additional_metrics.py
@@ -2,7 +2,7 @@ import numpy as np
 import sklearn.metrics
 
 
-def root_mean_squared_error(*args, **kwargs) -> float:
+def root_mean_squared_error(*args, **kwargs) -> np.float64:
     return sklearn.metrics.mean_squared_error(*args, **kwargs) ** 0.5
